requirements.txt
face_recognition>=1.3.0
osxphotos>=0.68.0
numpy>=1.24
Pillow>=10.0
pillow-heif>=0.13
tqdm>=4.66
